pyproject.toml
[project]
name = "fraud_detection"
version = "0.1.0"
description = "Portfolio fraud detection project"
authors = [
    {name = "Esosa Orumwese"}
]
license = {text = "MIT License"}
#readme = "README.md"
requires-python = ">=3.11,<3.13"
# Tell Poetry: “Look for packages under src/”
packages = [
  { include = "fraud_detection", from = "src" }
]
dependencies = [
    "requests (>=2.32.3,<3.0.0)",       # add using `poetry add <package-name>`
    "pyarrow (>=14.0.0,<22.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "mimesis (>=18.0.0,<19.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "polars (>=1.30.0,<2.0.0)",
    "geopandas (>=1.1.0,<2.0.0)",
    "rasterio (>=1.4.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "boto3 (==1.35.49)",
    "ydata-profiling (>=4.16.1,<5.0.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "xgboost (>=3.0.2,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "imbalanced-learn (>=0.13.0,<0.14.0)",
    "shap (>=0.47.2,<0.48.0)",
    "shapely (>=2.1.0,<3.0.0)",
    "pyproj (>=3.6.0,<4.0.0)",
    "referencing (>=0.35.0,<1.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "feast (>=0.41.0,<0.45.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "flask (==2.2.5)",
    "connexion[swagger-ui] (==2.14.2)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "scipy (>=1.14.0,<1.16.0)",
    "psutil (>=6.1.0,<7.0.0)",
]

[project.scripts]
engine-s0-run = "engine.cli.s0_foundations:main"
engine-s1-run = "engine.cli.s1_hurdle:main"
engine-s1-tile-index = "engine.cli.s1_tile_index:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]

[tool.poetry.group.dev.dependencies]    # add using `poetry add --dev <package-name>`
pytest = "^8.3.5"
detect-secrets = "^1.5.0"
pre-commit = "^4.2.0"
black = "^25.1.0"
ruff = "^0.11.9"
mypy = "^1.15.0"
coverage = "^7.8.0"
pytest-cov = "^6.1.1"
#checkov = "^3.2.427"
yamllint = "^1.37.1"
#great-expectations = "^1.4.5"
#pandas = ">=1.3,<2.2"
apache-airflow = {version = "3.0.2", extras = ["amazon", "celery", "postgres", "redis"]}


